Changed stepper so it's a little reusable. Still fixed length at 3, will update later if necessary
